requirements.txt
streamlit
openai>=1.0.0
spacy
fuzzywuzzy
python-docx
pdfplumber
pandas
